+++ examples/configs/recipes/llm/grpo-dapomath17k-qwen-30ba3b-megatron-zero-train-gen-kl.yaml
@@ -69,4 +69,4 @@ data:
 env:
   math:
     math_verify_impl: dapo_math_verify
-    truncate_solution_tail: false
+    dapo_long_cot: true
