+++ nemo_rl/models/generation/megatron/zero_train_gen_kl_patches/__init__.py
@@ -16,8 +16,10 @@
 
 from nemo_rl.models.generation.megatron.zero_train_gen_kl_patches.core_patches import (
     apply_log_softmax_determinism_patch,
+    apply_te_bik_attention_assert_skip_patch,
     apply_te_gemm_cublas_pinned_patch,
     restore_log_softmax_determinism_patch,
+    restore_te_bik_attention_assert_skip_patch,
     restore_te_gemm_cublas_pinned_patch,
 )
 from nemo_rl.models.generation.megatron.zero_train_gen_kl_patches.mamba_zero_kl_patches import (
@@ -34,10 +36,12 @@ __all__ = [
     "apply_log_softmax_determinism_patch",
     "apply_mamba_alignment_patch",
     "apply_moe_determinism_patches",
+    "apply_te_bik_attention_assert_skip_patch",
     "apply_te_gemm_cublas_pinned_patch",
     "policy_uses_mamba_layers",
     "restore_log_softmax_determinism_patch",
     "restore_mamba_alignment_patch",
     "restore_moe_determinism_patches",
+    "restore_te_bik_attention_assert_skip_patch",
     "restore_te_gemm_cublas_pinned_patch",
 ]
